Allow keeping more expressions.
