+++ src/low_res/swc_helpers.rs
@@ -1,6 +1,12 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
+use deno_ast::swc::ast::Expr;
 use deno_ast::swc::ast::Ident;
+use deno_ast::swc::ast::Lit;
+use deno_ast::swc::ast::MemberProp;
+use deno_ast::swc::ast::Prop;
+use deno_ast::swc::ast::PropName;
+use deno_ast::swc::ast::PropOrSpread;
 use deno_ast::swc::ast::ReturnStmt;
 use deno_ast::swc::ast::Stmt;
 use deno_ast::swc::ast::TsKeywordType;
@@ -99,4 +105,87 @@ pub fn source_range_to_range(
 ) -> Range {
   let position_range = PositionRange::from_source_range(range, text_info);
   Range::from_position_range(specifier.clone(), position_range)
+}
+
+pub fn is_expr_leavable(expr: &Expr) -> bool {
+  fn is_member_prop_leavable(n: &MemberProp) -> bool {
+    match n {
+      MemberProp::Ident(_) => true,
+      MemberProp::PrivateName(_) => false,
+      MemberProp::Computed(n) => is_expr_leavable(&n.expr),
+    }
+  }
+
+  match expr {
+    Expr::This(_) => true,
+    Expr::Array(n) => n.elems.iter().all(|elem| match elem {
+      Some(elem) => is_expr_leavable(&elem.expr),
+      None => true,
+    }),
+    Expr::Object(n) => n.props.iter().all(|prop| match prop {
+      PropOrSpread::Prop(prop) => match &**prop {
+        Prop::Shorthand(_) => true,
+        Prop::KeyValue(prop) => match &prop.key {
+          PropName::Ident(_) => is_expr_leavable(&prop.value),
+          PropName::Str(_) => is_expr_leavable(&prop.value),
+          PropName::Num(_) => is_expr_leavable(&prop.value),
+          PropName::Computed(c) => {
+            is_expr_leavable(&c.expr) && is_expr_leavable(&prop.value)
+          }
+          PropName::BigInt(_) => is_expr_leavable(&prop.value),
+        },
+        Prop::Assign(n) => is_expr_leavable(&n.value),
+        Prop::Getter(_) | Prop::Setter(_) | Prop::Method(_) => false,
+      },
+      PropOrSpread::Spread(n) => {
+        is_expr_leavable(&n.expr)
+      },
+    }),
+    Expr::Unary(n) => is_expr_leavable(&n.arg),
+    Expr::Update(n) => is_expr_leavable(&n.arg),
+    Expr::Bin(n) => is_expr_leavable(&n.left) && is_expr_leavable(&n.right),
+    Expr::Assign(_) | Expr::SuperProp(_) => false,
+    Expr::Cond(n) => {
+      is_expr_leavable(&n.test)
+        && is_expr_leavable(&n.cons)
+        && is_expr_leavable(&n.alt)
+    }
+    Expr::Member(n) => {
+      is_expr_leavable(&n.obj) && is_member_prop_leavable(&n.prop)
+    }
+    Expr::Ident(_) => true,
+    Expr::Call(_) | Expr::New(_) | Expr::Seq(_)  => false,
+    Expr::Lit(n) => match n {
+      Lit::Str(_)
+      | Lit::Bool(_)
+      | Lit::Null(_)
+      | Lit::Num(_)
+      | Lit::BigInt(_)
+      | Lit::Regex(_) => true,
+      Lit::JSXText(_) => false,
+    },
+    Expr::Await(n) => is_expr_leavable(&n.arg),
+    Expr::Paren(n) => is_expr_leavable(&n.expr),
+    Expr::TsTypeAssertion(_) | Expr::TsAs(_) => false,
+    Expr::TsConstAssertion(n) => is_expr_leavable(&n.expr),
+    Expr::TsNonNull(n) => is_expr_leavable(&n.expr),
+    Expr::Tpl(_)
+    | Expr::Fn(_)
+    | Expr::TaggedTpl(_)
+    | Expr::Arrow(_)
+    | Expr::Class(_)
+    | Expr::Yield(_)
+    | Expr::MetaProp(_)
+    | Expr::JSXMember(_)
+    | Expr::JSXNamespacedName(_)
+    | Expr::JSXEmpty(_)
+    | Expr::JSXElement(_)
+    | Expr::JSXFragment(_)
+    | Expr::TsInstantiation(_)
+    | Expr::TsSatisfies(_)
+    | Expr::PrivateName(_)
+    // todo: probably could analyze this more
+    | Expr::OptChain(_)
+    | Expr::Invalid(_) => false,
+  }
 }
