+++ tests/specs/graph/JsrSpecifiers_LowRes_InitIdentsAndMembers.txt
@@ -1,0 +1,149 @@
+# https://jsr.io/@scope/a/meta.json
+{"versions": { "1.0.0": {} } }
+
+# https://jsr.io/@scope/a/1.0.0_meta.json
+{ "exports": { ".": "./mod.ts" } }
+
+# https://jsr.io/@scope/a/1.0.0/mod.ts
+class BaseHandler {
+}
+
+class Public {
+}
+
+class Private {}
+
+namespace Test {
+  export class A {}
+  // should not include this
+  export class B {}
+}
+
+export const handlers = {
+  BaseHandler,
+  Private: Public,
+  C: Test.A,
+};
+
+# https://jsr.io/@scope/b/meta.json
+{"versions": { "1.0.0": {} } }
+
+# https://jsr.io/@scope/b/1.0.0_meta.json
+{ "exports": { ".": "./mod.ts" } }
+
+# https://jsr.io/@scope/b/1.0.0/mod.ts
+class BaseHandler {
+}
+
+namespace Test {
+  export class A {}
+  // should not include this
+  export class B {}
+}
+
+export function test(
+  a = BaseHandler,
+  C = Test.A,
+) {
+}
+
+# mod.ts
+import 'jsr:@scope/a'
+import 'jsr:@scope/b'
+
+# output
+{
+  "roots": [
+    "file:///mod.ts"
+  ],
+  "modules": [
+    {
+      "kind": "esm",
+      "dependencies": [
+        {
+          "specifier": "jsr:@scope/a",
+          "code": {
+            "specifier": "jsr:@scope/a",
+            "span": {
+              "start": {
+                "line": 0,
+                "character": 7
+              },
+              "end": {
+                "line": 0,
+                "character": 21
+              }
+            }
+          }
+        },
+        {
+          "specifier": "jsr:@scope/b",
+          "code": {
+            "specifier": "jsr:@scope/b",
+            "span": {
+              "start": {
+                "line": 1,
+                "character": 7
+              },
+              "end": {
+                "line": 1,
+                "character": 21
+              }
+            }
+          }
+        }
+      ],
+      "size": 44,
+      "mediaType": "TypeScript",
+      "specifier": "file:///mod.ts"
+    },
+    {
+      "kind": "esm",
+      "size": 224,
+      "mediaType": "TypeScript",
+      "specifier": "https://jsr.io/@scope/a/1.0.0/mod.ts"
+    },
+    {
+      "kind": "esm",
+      "size": 173,
+      "mediaType": "TypeScript",
+      "specifier": "https://jsr.io/@scope/b/1.0.0/mod.ts"
+    }
+  ],
+  "redirects": {
+    "jsr:@scope/a": "https://jsr.io/@scope/a/1.0.0/mod.ts",
+    "jsr:@scope/b": "https://jsr.io/@scope/b/1.0.0/mod.ts"
+  },
+  "packages": {
+    "@scope/a": "@scope/a@1.0.0",
+    "@scope/b": "@scope/b@1.0.0"
+  }
+}
+
+Low res https://jsr.io/@scope/a/1.0.0/mod.ts:
+  {}
+  class BaseHandler {
+  }
+  class Public {
+  }
+  module Test {
+    export class A {
+    }
+  }
+  export const handlers = {
+    BaseHandler,
+    Private: Public,
+    C: Test.A
+  };
+
+Low res https://jsr.io/@scope/b/1.0.0/mod.ts:
+  {}
+  class BaseHandler {
+  }
+  module Test {
+    export class A {
+    }
+  }
+  export function test(a = BaseHandler, C = Test.A): void {
+    return {} as any;
+  }
