+++ tests/specs/graph/JsrSpecifiers_LowRes_RefObjType.txt
@@ -5,11 +5,18 @@
 { "exports": { ".": "./mod.ts" } }
 
 # https://jsr.io/@scope/a/1.0.0/mod.ts
-class Class {}
+export const STATUS_CODE = {
+  /** RFC 7231, 6.2.1 */
+  Continue: 100,
+  /** RFC 7231, 6.2.2 */
+  SwitchingProtocols: 101,
+  /** RFC 2518, 10.1 */
+  Processing: 102,
+  /** RFC 8297 **/
+  EarlyHints: 103,
+}
 
-export default {
-  test: new Class(),
-};
+export type Status = typeof STATUS_CODE.Continue;
 
 # mod.ts
 import 'jsr:@scope/a'
@@ -46,7 +53,7 @@ import 'jsr:@scope/a'
     },
     {
       "kind": "esm",
-      "size": 57,
+      "size": 257,
       "mediaType": "TypeScript",
       "specifier": "https://jsr.io/@scope/a/1.0.0/mod.ts"
     }
@@ -60,5 +67,11 @@ import 'jsr:@scope/a'
 }
 
 Low res https://jsr.io/@scope/a/1.0.0/mod.ts:
-  Default export expression was too complex. Extract it out to a variable and add an explicit type.
-      at https://jsr.io/@scope/a/1.0.0/mod.ts:3:1
+  {}
+  export const STATUS_CODE = {
+    /** RFC 7231, 6.2.1 */ Continue: 100,
+    /** RFC 7231, 6.2.2 */ SwitchingProtocols: 101,
+    /** RFC 2518, 10.1 */ Processing: 102,
+    /** RFC 8297 **/ EarlyHints: 103
+  };
+  export type Status = typeof STATUS_CODE.Continue;
