+++ tests/specs/symbols/ExportDefault04.txt
@@ -34,12 +34,49 @@ file:///mod.ts: EsmModuleInfo {
                     flags: 0,
                 },
             ],
+            child_ids: {
+                1,
+            },
+            exports: {
+                "default": 1,
+            },
+            members: {},
+        },
+        1: Symbol {
+            module_id: ModuleId(
+                0,
+            ),
+            symbol_id: 1,
+            parent_id: Some(
+                0,
+            ),
+            decls: [
+                SymbolDecl {
+                    kind: Definition(
+                        SymbolNode(
+                            "export default 1 + 1;",
+                        ),
+                    ),
+                    range: SourceRange {
+                        start: SourcePos(
+                            0,
+                        ),
+                        end: SourcePos(
+                            21,
+                        ),
+                    },
+                    flags: 0,
+                },
+            ],
             child_ids: {},
             exports: {},
             members: {},
         },
     },
 }
+== export definitions ==
+[default]: file:///mod.ts:0..21
+  export default 1 + 1;
 
 # diagnostics
 [
